Midterms ver 3
Everything works... except I can't repeat the Pop and Dequeue options without throwing an ArrayExceptionOutofBounds100 at me. I have no idea why. The goal is this: every time I push/queue a number, I should put it in an array. I should be able to enter one input at a time but can repeat the Push option many times. When I choose the Pop option, it must remove the last value inputted. I can use Pop until there are no values left in the array, in which case it tells the user that there are no values in the stack. It's the same with Queue/Dequeue, except the removed value is always the first value inputted. If I choose dequeue again, it should remove the second earliest value inputted, and so on and so forth.